docs/archive/spikes/spike3_llm_extraction/requirements.txt
anthropic>=0.36.0
pydantic>=2.0.0
python-dotenv>=1.0.0
